requirements.txt
flask
flask-cors
google-cloud-aiplatform
python-dotenv
google-generativeai
